+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 [Unreleased](https://github.com/jshwi/docsig/compare/v0.33.1...HEAD)
 ------------------------------------------------------------------------
+### Fix
+- Relax `arcon` version constraint
 
 [0.33.1](https://github.com/jshwi/docsig/releases/tag/v0.33.1) - 2022-12-24
 ------------------------------------------------------------------------