+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 [Unreleased](https://github.com/jshwi/docsig/compare/v0.33.2...HEAD)
 ------------------------------------------------------------------------
+### Added
+- Add support for Python 3.10
 
 [0.33.2](https://github.com/jshwi/docsig/releases/tag/v0.33.2) - 2023-01-05
 ------------------------------------------------------------------------